refactor: extract AdminerService, fix namespace bug, improve CSS
- Move Adminer logic from resources/adminer/index.php into
  AdminerService + adminer_object.php (global namespace required)
- Fix adminer_object() being scoped to App\Services namespace
  instead of global, which broke auto-login in iframe
- Vite-built CSS with vendor base + overrides (adminer.css)
- Full-width breadcrumb, hide logo/logout/lang/server link
- Modernized inputs, buttons, table headers
