+++ phpstan.neon
@@ -13,6 +13,9 @@ parameters:
     checkModelProperties: true
     noEnvCallsOutsideOfConfig: true
 
+    excludePaths:
+        - app/Services/adminer_object.php
+
     ignoreErrors:
         - '#Call to an undefined method Illuminate\\Database\\Eloquent\\Builder.*::allowedFilters\(\)#'
         - '#should return .* but returns .* \$this\(.*\)#'
